+++ src/RcppExports.cpp
@@ -91,6 +91,20 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// cpp_lazy_distance_sd
+double cpp_lazy_distance_sd(Rcpp::NumericMatrix left_mat, Rcpp::NumericMatrix right_mat, std::string metric, Rcpp::Nullable<Rcpp::NumericMatrix> inv_cov);
+RcppExport SEXP _couplr_cpp_lazy_distance_sd(SEXP left_matSEXP, SEXP right_matSEXP, SEXP metricSEXP, SEXP inv_covSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< Rcpp::NumericMatrix >::type left_mat(left_matSEXP);
+    Rcpp::traits::input_parameter< Rcpp::NumericMatrix >::type right_mat(right_matSEXP);
+    Rcpp::traits::input_parameter< std::string >::type metric(metricSEXP);
+    Rcpp::traits::input_parameter< Rcpp::Nullable<Rcpp::NumericMatrix> >::type inv_cov(inv_covSEXP);
+    rcpp_result_gen = Rcpp::wrap(cpp_lazy_distance_sd(left_mat, right_mat, metric, inv_cov));
+    return rcpp_result_gen;
+END_RCPP
+}
 // cpp_lap_solve_auction_lazy
 Rcpp::List cpp_lap_solve_auction_lazy(Rcpp::NumericMatrix left_mat, Rcpp::NumericMatrix right_mat, std::string metric, Rcpp::Nullable<Rcpp::NumericMatrix> inv_cov, double max_distance, Rcpp::List calipers, Rcpp::CharacterVector var_names, bool maximize, Rcpp::Nullable<double> eps);
 RcppExport SEXP _couplr_cpp_lap_solve_auction_lazy(SEXP left_matSEXP, SEXP right_matSEXP, SEXP metricSEXP, SEXP inv_covSEXP, SEXP max_distanceSEXP, SEXP calipersSEXP, SEXP var_namesSEXP, SEXP maximizeSEXP, SEXP epsSEXP) {
@@ -557,6 +571,30 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// lap_full_match_implicit
+Rcpp::List lap_full_match_implicit(Rcpp::NumericMatrix left_mat, Rcpp::NumericMatrix right_mat, std::string distance, Rcpp::Nullable<Rcpp::NumericMatrix> inv_cov, double max_distance, Rcpp::List calipers, Rcpp::CharacterVector vars, double min_controls, double max_controls, double keep_per_row, double width, double tol, double max_rounds, bool certify);
+RcppExport SEXP _couplr_lap_full_match_implicit(SEXP left_matSEXP, SEXP right_matSEXP, SEXP distanceSEXP, SEXP inv_covSEXP, SEXP max_distanceSEXP, SEXP calipersSEXP, SEXP varsSEXP, SEXP min_controlsSEXP, SEXP max_controlsSEXP, SEXP keep_per_rowSEXP, SEXP widthSEXP, SEXP tolSEXP, SEXP max_roundsSEXP, SEXP certifySEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< Rcpp::NumericMatrix >::type left_mat(left_matSEXP);
+    Rcpp::traits::input_parameter< Rcpp::NumericMatrix >::type right_mat(right_matSEXP);
+    Rcpp::traits::input_parameter< std::string >::type distance(distanceSEXP);
+    Rcpp::traits::input_parameter< Rcpp::Nullable<Rcpp::NumericMatrix> >::type inv_cov(inv_covSEXP);
+    Rcpp::traits::input_parameter< double >::type max_distance(max_distanceSEXP);
+    Rcpp::traits::input_parameter< Rcpp::List >::type calipers(calipersSEXP);
+    Rcpp::traits::input_parameter< Rcpp::CharacterVector >::type vars(varsSEXP);
+    Rcpp::traits::input_parameter< double >::type min_controls(min_controlsSEXP);
+    Rcpp::traits::input_parameter< double >::type max_controls(max_controlsSEXP);
+    Rcpp::traits::input_parameter< double >::type keep_per_row(keep_per_rowSEXP);
+    Rcpp::traits::input_parameter< double >::type width(widthSEXP);
+    Rcpp::traits::input_parameter< double >::type tol(tolSEXP);
+    Rcpp::traits::input_parameter< double >::type max_rounds(max_roundsSEXP);
+    Rcpp::traits::input_parameter< bool >::type certify(certifySEXP);
+    rcpp_result_gen = Rcpp::wrap(lap_full_match_implicit(left_mat, right_mat, distance, inv_cov, max_distance, calipers, vars, min_controls, max_controls, keep_per_row, width, tol, max_rounds, certify));
+    return rcpp_result_gen;
+END_RCPP
+}
 // lap_flow_compile_couples
 Rcpp::List lap_flow_compile_couples(std::string design, double n_rows, double n_cols, double ratio);
 RcppExport SEXP _couplr_lap_flow_compile_couples(SEXP designSEXP, SEXP n_rowsSEXP, SEXP n_colsSEXP, SEXP ratioSEXP) {
@@ -936,6 +974,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_couplr_lap_solve_jv", (DL_FUNC) &_couplr_lap_solve_jv, 2},
     {"_couplr_cpp_lap_solve_jv_lazy", (DL_FUNC) &_couplr_cpp_lap_solve_jv_lazy, 8},
     {"_couplr_cpp_lazy_pair_distances", (DL_FUNC) &_couplr_cpp_lazy_pair_distances, 6},
+    {"_couplr_cpp_lazy_distance_sd", (DL_FUNC) &_couplr_cpp_lazy_distance_sd, 4},
     {"_couplr_cpp_lap_solve_auction_lazy", (DL_FUNC) &_couplr_cpp_lap_solve_auction_lazy, 9},
     {"_couplr_lap_certify_dense", (DL_FUNC) &_couplr_lap_certify_dense, 7},
     {"_couplr_lap_certify_lazy", (DL_FUNC) &_couplr_lap_certify_lazy, 13},
@@ -969,6 +1008,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_couplr_lap_flow_solve", (DL_FUNC) &_couplr_lap_flow_solve, 14},
     {"_couplr_lap_flow_certify", (DL_FUNC) &_couplr_lap_flow_certify, 10},
     {"_couplr_lap_flow_compile_full_match", (DL_FUNC) &_couplr_lap_flow_compile_full_match, 3},
+    {"_couplr_lap_full_match_implicit", (DL_FUNC) &_couplr_lap_full_match_implicit, 14},
     {"_couplr_lap_flow_compile_couples", (DL_FUNC) &_couplr_lap_flow_compile_couples, 4},
     {"_couplr_lap_flow_trace_assignment", (DL_FUNC) &_couplr_lap_flow_trace_assignment, 2},
     {"_couplr_lap_flow_trace_push_relabel", (DL_FUNC) &_couplr_lap_flow_trace_push_relabel, 2},